refactoring

+++ src/commands/clone.rs
@@ -1,7 +1,7 @@
-use super::{construct_path, save_config, upsert_metadata, Config};
+use super::{construct_local_path, create_metadir, save_config, upsert_metadata, Config};
 use crate::db;
+use crate::files::download;
 use crate::files::list_folder::Entry;
-use crate::files::{download, list_folder};
 use crate::ignore::{parce_ignore, IGNORE_FILE};
 use tokio::fs;
 
@@ -9,67 +9,51 @@ use std::path::Path;
 
 pub async fn clone(
     remote_path: &str,
-    local_path: &Path,
+    local_root: &Path,
     token: &str,
 ) -> Result<(), Box<dyn std::error::Error>> {
-    let curr_dir = std::env::current_dir()?;
-    let is_curr = curr_dir == local_path;
     let ignore_filter = parce_ignore().await?;
 
-    fs::create_dir_all(local_path).await?;
-    let conn = db::connect(local_path)?;
-
-    let mut dirs = vec![remote_path.to_owned()];
-    let mut files: Vec<String> = vec![];
+    fs::create_dir_all(local_root).await?;
+    let conn = db::connect(local_root)?;
 
     let config = Config {
         remote_path: remote_path.to_owned(),
     };
 
-    // Download process is intentionally non-parallel
-    while dirs.len() > 0 || files.len() > 0 {
-        for file in files.iter() {
-            if !ignore_filter.is_ignored(file) {
-                download_file(file, &config, local_path, &conn, token).await?;
-            }
-        }
+    println!(
+        "Cloning {} into {} ...",
+        remote_path,
+        local_root.to_string_lossy()
+    );
 
-        let mut new_dirs = vec![];
-        let mut new_files = vec![];
+    visit_all_files_and_dirs!(
+        vec![remote_path.to_owned()],
+        Vec::<String>::new(),
+        &ignore_filter,
+        &config,
+        local_root,
+        &conn,
+        token,
+        download_file,
+        create_dirs
+    );
 
-        for dir in dirs.iter() {
-            let entries = read_dir(dir, &config, local_path, token).await?;
-
-            for entry in entries.into_iter() {
-                match entry {
-                    Entry::File(file_info) => {
-                        if let Some(name) = file_info.path_display {
-                            new_files.push(name);
-                        }
-                    }
-                    Entry::Folder { path_display, .. } => {
-                        if let Some(name) = path_display {
-                            new_dirs.push(name);
-                        }
-                    }
-                    _ => (),
-                }
-            }
-        }
-
-        dirs = new_dirs;
-        files = new_files;
-    }
-
+    let curr_dir = std::env::current_dir()?;
+    let is_curr = std::fs::canonicalize(&curr_dir)? == std::fs::canonicalize(local_root)?;
     if !is_curr {
         let mut ignore_file = curr_dir.clone();
         ignore_file.push(IGNORE_FILE);
-        let mut ignore_file_dst = local_path.to_owned();
+        let mut ignore_file_dst = local_root.to_owned();
         ignore_file_dst.push(IGNORE_FILE);
-        fs::copy(ignore_file, ignore_file_dst).await?;
+        if ignore_file.exists() {
+            fs::copy(ignore_file, ignore_file_dst).await?;
+        }
     }
 
-    save_config(&config, local_path).await?;
+    save_config(&config, local_root).await?;
+
+    println!("done.");
     Ok(())
 }
 
@@ -80,27 +64,27 @@ async fn download_file(
     conn: &rusqlite::Connection,
     token: &str,
 ) -> Result<(), Box<dyn std::error::Error>> {
-    let local_file = construct_path(remote_path, config, local_root);
+    let local_file = construct_local_path(remote_path, config, local_root);
+    println!(
+        "Found file {}. Downloading to {} ...",
+        remote_path,
+        local_file.to_string_lossy()
+    );
     let (info, data) = download::download(remote_path, token).await?;
     upsert_metadata(local_root, conn, config, info, &data).await?;
     fs::write(local_file, data).await?;
     Ok(())
 }
 
-async fn read_dir(
-    remote_path: &str,
+async fn create_dirs(
+    remote_dir: &str,
     config: &Config,
     local_root: &Path,
-    token: &str,
-) -> Result<Vec<Entry>, Box<dyn std::error::Error>> {
-    let local_dir = construct_path(remote_path, config, local_root);
+    _conn: &rusqlite::Connection,
+) -> Result<(), Box<dyn std::error::Error>> {
+    let local_dir = construct_local_path(remote_dir, config, local_root);
+    println!("Creating directory {} ...", local_dir.to_string_lossy());
     fs::create_dir_all(local_dir).await?;
-    let entries = list_folder::list_folder(
-        remote_path,
-        token,
-        #[cfg(test)]
-        200,
-    )
-    .await?;
-    Ok(entries)
+    create_metadir(remote_dir, config, local_root).await?;
+    Ok(())
 }
